+++ common/src/main/java/dev/aika/smsn/client/gui/ComponentBuilder.java
@@ -7,13 +7,10 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.experimental.Accessors;
 import me.shedaniel.clothconfig2.api.ConfigEntryBuilder;
-import net.fabricmc.api.EnvType;
-import net.fabricmc.api.Environment;
 
 import java.lang.reflect.Field;
 
 @Accessors(chain = true)
-@Environment(EnvType.CLIENT)
 public class ComponentBuilder {
     @Getter
     protected final ConfigEntryBuilder entryBuilder;