+++ common/src/main/java/dev/aika/smsn/client/gui/components/AbstractComponentBuilder.java
@@ -10,8 +10,6 @@ import me.shedaniel.clothconfig2.api.AbstractConfigListEntry;
 import me.shedaniel.clothconfig2.api.ConfigEntryBuilder;
 import me.shedaniel.clothconfig2.impl.builders.AbstractFieldBuilder;
 import me.shedaniel.clothconfig2.impl.builders.FieldBuilder;
-import net.fabricmc.api.EnvType;
-import net.fabricmc.api.Environment;
 import net.minecraft.network.chat.Component;
 import org.slf4j.Logger;
 import org.slf4j.Marker;
@@ -21,7 +19,6 @@ import java.lang.reflect.Field;
 import java.lang.reflect.Modifier;
 
 @Accessors(chain = true)
-@Environment(EnvType.CLIENT)
 public abstract class AbstractComponentBuilder<T> {
     private static final Logger log = SMSN.LOGGER;
     private static final Marker marker = MarkerFactory.getMarker("AbstractComponentBuilder");