+++ neoforge/src/main/java/dev/aika/smsn/neoforge/mixin/xxrexraptorxx/allthecompatibility/EventsMixin.java
@@ -1,4 +1,4 @@
-package dev.aika.smsn.neoforge.mixin.additionalstructures;
+package dev.aika.smsn.neoforge.mixin.xxrexraptorxx.allthecompatibility;
 
 import dev.aika.smsn.SMSN;
 import net.minecraft.world.entity.player.Player;
@@ -8,7 +8,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-import xxrexraptorxx.additionalstructures.utils.Events;
+import xxrexraptorxx.allthecompatibility.utils.Events;
 
 import java.net.URL;
 
@@ -16,11 +16,16 @@ import java.net.URL;
 public abstract class EventsMixin {
     @Inject(method = "SupporterCheck", at = @At("HEAD"), cancellable = true)
     private static void SupporterCheck(URL url, Player player, CallbackInfoReturnable<Boolean> cir) {
-        if (SMSN.CONFIG.isAdditionalStructuresSupporterCheck()) cir.setReturnValue(false);
+        if (SMSN.CONFIG.isXxRexRaptorxXModsSupporterCheck()) cir.setReturnValue(false);
     }
 
     @Inject(method = "SupporterRewards", at = @At("HEAD"), cancellable = true)
     private static void SupporterRewards(PlayerEvent.PlayerLoggedInEvent event, CallbackInfo ci) {
-        if (SMSN.CONFIG.isAdditionalStructuresSupporterCheck()) ci.cancel();
+        if (SMSN.CONFIG.isXxRexRaptorxXModsSupporterCheck()) ci.cancel();
+    }
+
+    @Inject(method = "onPlayerLogin", at = @At("HEAD"), cancellable = true)
+    private static void onPlayerLogin(PlayerEvent.PlayerLoggedInEvent event, CallbackInfo ci) {
+        if (!SMSN.CONFIG.isStopModRepostsMessages()) ci.cancel();
     }
 }